Add new exception and move existing into exceptions folder.
