+++ src/QueueFailure.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types=1);
+
+namespace MaxBrokman\SafeQueue;
+
+/**
+ * Interface QueueMustStop used in testing to force a stop.
+ */
+interface QueueFailure
+{
+}
